+++ railties/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Rails 4.0.0 (unreleased) ##
 
+*   Load all environments available in `config.paths["config/environments"]`. *Piotr Sarnacki*
+
 *   The application generator generates `public/humans.txt` with some basic data. *Paul Campbell*
 
 *   Add `config.queue_consumer` to allow the default consumer to be configurable. *Carlos Antonio da Silva*